Update AuthenticationStateExtensions.cs
Hi, instead of using ApiKeys did you try capturing the SessionId as a claim then passing it by-header when calling the API? It seems to work as well for me.

However even without this change I was getting System.InvalidOperationException: ''RemoteNavigationManager' has not been initialized.' here `navigationManager.BaseUri;`. The base URI could just be put in a config setting I guess.

Thanks for this project, it's given me a few ideas for my apps.